@@ -1,7 +1,7 @@
 package com.vsa.steampartyfinder.ui.adapter.players
 
 import android.graphics.Color
-import android.support.v7.widget.RecyclerView
+import androidx.recyclerview.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
