@@ -2,8 +2,8 @@ package com.vsa.steampartyfinder.ui.base
 
 import android.app.ProgressDialog
 import android.os.Bundle
-import android.support.design.widget.Snackbar
-import android.support.v7.app.AppCompatActivity
+import com.google.android.material.snackbar.Snackbar
+import androidx.appcompat.app.AppCompatActivity
 import android.view.View
 import com.vsa.steampartyfinder.R
 
