@@ -2,8 +2,8 @@ package com.vsa.steampartyfinder.ui.games.filter
 
 import android.animation.Animator
 import android.os.Bundle
-import android.support.v4.app.DialogFragment
-import android.support.v4.app.FragmentManager
+import androidx.fragment.app.DialogFragment
+import androidx.fragment.app.FragmentManager
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
